Refactor auth codepath
